sync: init
This adds the initial specification for sync conduits, some helper
classes for implementing sync conduits (both on the `palmrs-sync` side
and the actual conduit binary side), and two new (incomplete) binaries:

- `palmrs-sync-dbgconduit` is a debug runner for sync conduits, allowing
  for running them outside of a full `palmrs-sync` (which is yet to be
  implemented anyway)
- `palmrs-conduit-todotxt` is a Palm OS "Tasks" / "To Do" <-> todo.txt
  format sync conduit

The primary focus for the next little while is going to be getting Palm
OS database *creation and modification* going, with the aim of getting
`palmrs-conduit-todotxt` actually working - reading the tasks out of an
existing Palm OS "Tasks" database and writing a todo.txt file, and then
creating a new Palm OS "Tasks" database from a todo.txt file, such that
I can then throw that database at Palm Desktop on a Windows machine and
have it sync it to my handheld.

We'll see!